crypto: Ai generated audit feedback (#3536)
Some small cleanups:

- Rename kbkdf 'salt' to 'label' to match the algorithm's official
definition
- Add some comments
- Set xts_aes_256's tweak size to u64 everywhere to match the bcrypt
restriction, so we don't accidentally depend on a larger tweak in the
future

+++ support/crypto/src/xts_aes_256/ossl.rs
@@ -54,8 +54,8 @@ impl XtsAes256Inner {
 }
 
 impl XtsAes256EncCtxInner<'_> {
-    pub fn cipher(&mut self, tweak: u128, data: &mut [u8]) -> Result<(), XtsAes256Error> {
-        let iv = tweak.to_le_bytes();
+    pub fn cipher(&mut self, tweak: u64, data: &mut [u8]) -> Result<(), XtsAes256Error> {
+        let iv = (tweak as u128).to_le_bytes();
         self.ctx
             .encrypt_init(None, None, Some(&iv))
             .map_err(|e| err(e, "encryption"))?;
@@ -67,8 +67,8 @@ impl XtsAes256EncCtxInner<'_> {
 }
 
 impl XtsAes256DecCtxInner<'_> {
-    pub fn cipher(&mut self, tweak: u128, data: &mut [u8]) -> Result<(), XtsAes256Error> {
-        let iv = tweak.to_le_bytes();
+    pub fn cipher(&mut self, tweak: u64, data: &mut [u8]) -> Result<(), XtsAes256Error> {
+        let iv = (tweak as u128).to_le_bytes();
         self.ctx
             .decrypt_init(None, None, Some(&iv))
             .map_err(|e| err(e, "decryption"))?;
